chore: moved 'insanity' inside of '/toys'
