+++ fresh.gen.ts
@@ -3,8 +3,8 @@
 // This file is automatically updated during development when running `dev.ts`.
 
 import config from "./deno.json" assert { type: "json" };
-import * as $0 from "./routes/[blog]/this.tsx";
-import * as $1 from "./routes/[toys]/index.tsx";
+import * as $0 from "./routes/[toys]/index.tsx";
+import * as $1 from "./routes/[toys]/insanity.tsx";
 import * as $2 from "./routes/[toys]/spinners.tsx";
 import * as $3 from "./routes/_app.tsx";
 import * as $4 from "./routes/api/joke.ts";
@@ -17,8 +17,8 @@ import * as $$3 from "./islands/ThemeSwitcher.tsx";
 
 const manifest = {
   routes: {
-    "./routes/[blog]/this.tsx": $0,
-    "./routes/[toys]/index.tsx": $1,
+    "./routes/[toys]/index.tsx": $0,
+    "./routes/[toys]/insanity.tsx": $1,
     "./routes/[toys]/spinners.tsx": $2,
     "./routes/_app.tsx": $3,
     "./routes/api/joke.ts": $4,
